graph_traversal.py
Tower of Hanoi solver: Solves the Tower of Hanoi problem by representing the puzzle as a graph and using Depth First Search and A* Search. Allows the user to input the initial state of the puzzle, number of disks. and computes minimum steps to solve the problem. Outputs the steps (path) to solve the problem.